add workflow

+++ index.sh
@@ -1,25 +1,44 @@
 #!/bin/bash
 
-# create AMI NGINX
-# cd ./packer
-# export AWS_ACCESS_KEY_ID=$1
-# export AWS_SECRET_ACCESS_KEY=$2
-# packer build index.json
-# cd ../
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
+echo $3
+if [[ "$3" ]];
+then
+    AMI_ID=$(jq -r '.builds[-1].artifact_id' ./packer/manifest.json | cut -d ":" -f2)
+    echo $AMI_ID
+    cd ./terraform/ec2
 
-# print created AMI ID
-AMI_ID=$(jq -r '.builds[-1].artifact_id' ./packer/manifest.json | cut -d ":" -f2)
-echo $AMI_ID
+    terraform destroy -var="access_key=$1" \
+        -var="secret_key=$2" \
+        -var="ami_id=$AMI_ID" \
+        -var-file=variables.tfvars
 
-#create EC2 Instance
-cd ./terraform/ec2
-# terraform init -upgrade
-terraform destroy -var="access_key=$1" \
-    -var="secret_key=$2" \
-    -var="ami_id=$AMI_ID" \
-    -var-file=variables.tfvars
-cd ../../ansible/
+    cd ../../
+else
+    # create AMI NGINX
+    # cd ./packer
+    # export AWS_ACCESS_KEY_ID=$1
+    # export AWS_SECRET_ACCESS_KEY=$2
+    # packer build index.json
+    # cd ../
 
-# ansible-playbook -i inventories nginx.yml
+    # print created AMI ID
+    AMI_ID=$(jq -r '.builds[-1].artifact_id' ./packer/manifest.json | cut -d ":" -f2)
+    echo ami $AMI_ID
 
-cd ../
+    #create EC2 Instance
+    cd ./terraform/ec2
+    terraform init -upgrade
+    terraform apply -var="access_key=$1" \
+        -var="secret_key=$2" \
+        -var="ami_id=$AMI_ID" \
+        -var-file=variables.tfvars
+
+    terraform output aws_instance >> "$SCRIPT_DIR/ansible/inventories"
+
+    cd ../../ansible/
+
+    ansible-playbook -i inventories nginx-container.yml
+
+    cd ../
+fi
